centralizar tipos y TIME_UNIT en common.ts
- Se creó src/Components/Algorithms/common.ts
- Se movieron las interfaces Process, ExecutionStep, ProcessResult y QueueSnapshot desde SRTF.ts
- Se exportó TIME_UNIT como configuración global de la UI
- Se ajustaron imports en SRTF.ts y SRTFSimulator.tsx para usar el nuevo archivo
